requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
opencv-python==4.8.0.74
pillow==10.0.0
matplotlib==3.7.2
streamlit==1.25.0
scipy==1.11.1
